+++ aws-shared/src/main/java/com/kdgregory/logging/aws/cloudwatch/CloudWatchLogWriter.java
@@ -108,6 +108,13 @@ extends AbstractLogWriter<CloudWatchWriterConfig,CloudWatchWriterStatistics,AWSL
             && (numMessages <= CloudWatchConstants.MAX_BATCH_COUNT);
     }
 
+
+    @Override
+    protected void stopAWSClient()
+    {
+        client.shutdown();
+    }
+
 //----------------------------------------------------------------------------
 //  Internals
 //----------------------------------------------------------------------------