+++ aws-shared/src/main/java/com/kdgregory/logging/aws/internal/AbstractLogWriter.java
@@ -121,17 +121,6 @@ implements LogWriter
 
 
     @Override
-    public void stop()
-    {
-        shutdownTime = System.currentTimeMillis() + config.batchDelay;
-        if (dispatchThread != null)
-        {
-            dispatchThread.interrupt();
-        }
-    }
-
-
-    @Override
     public void setDiscardThreshold(int value)
     {
         messageQueue.setDiscardThreshold(value);
@@ -142,6 +131,17 @@ implements LogWriter
     public void setDiscardAction(DiscardAction value)
     {
         messageQueue.setDiscardAction(value);
+    }
+
+
+    @Override
+    public void stop()
+    {
+        shutdownTime = System.currentTimeMillis() + config.batchDelay;
+        if (dispatchThread != null)
+        {
+            dispatchThread.interrupt();
+        }
     }
 
 //----------------------------------------------------------------------------
@@ -181,6 +181,10 @@ implements LogWriter
                 requeueMessages(failures);
             }
         } while (keepRunning());
+
+        stopAWSClient();
+        logger.debug("stopping log-writer on thread " + Thread.currentThread().getName() 
+                     + " (#" + Thread.currentThread().getId() + ")");
     }
 
 //----------------------------------------------------------------------------
@@ -215,6 +219,50 @@ implements LogWriter
     {
         return shutdownTime > System.currentTimeMillis()
             || ! messageQueue.isEmpty();
+    }
+
+
+    /**
+     *  Attempts to read a list of messages from the queue. Will wait "forever"
+     *  (or until shutdown) for the first message, then read as many messages
+     *  as possible within the batch delay.
+     *  <p>
+     *  For each message, the subclass is called to determine the effective size
+     *  of the message, and whether the aggregate batch size is within the range
+     *  accepted by the service.
+     */
+    protected List<LogMessage> buildBatch()
+    {
+        // presizing to a small-but-possible size to avoid repeated resizes
+        List<LogMessage> batch = new ArrayList<LogMessage>(512);
+
+        // we'll wait "forever" unless there's a shutdown timestamp in effect
+        LogMessage message = waitForMessage(shutdownTime);
+        if (message == null)
+            return batch;
+
+        long batchTimeout = System.currentTimeMillis() + config.batchDelay;
+        int batchBytes = 0;
+        int batchMsgs = 0;
+        while (message != null)
+        {
+            batchBytes += effectiveSize(message);
+            batchMsgs++;
+
+            // if this message would exceed the batch limits, push it back onto the queue
+            // the first message must never break this rule -- and shouldn't, as long as
+            // appender checks size
+            if (! withinServiceLimits(batchBytes, batchMsgs))
+            {
+                messageQueue.requeue(message);
+                break;
+            }
+
+            batch.add(message);
+            message = waitForMessage(batchTimeout);
+        }
+
+        return batch;
     }
 
 
@@ -275,6 +323,14 @@ implements LogWriter
      */
     protected abstract boolean withinServiceLimits(int batchBytes, int numMessages);
 
+
+    /**
+     *  This is called when the logwriter is stopped, to explicitly close the
+     *  AWS service client. It must be implemented by the subclass because we
+     *  don't know the actual type and there's no abstract super-interface.
+     */
+    protected abstract void stopAWSClient();
+
 //----------------------------------------------------------------------------
 //  Subclass helpers
 //----------------------------------------------------------------------------
@@ -287,49 +343,5 @@ implements LogWriter
     {
         logger.error(message, exception);
         stats.setLastError(message, exception);
-    }
-
-
-    /**
-     *  Attempts to read a list of messages from the queue. Will wait "forever"
-     *  (or until shutdown) for the first message, then read as many messages
-     *  as possible within the batch delay.
-     *  <p>
-     *  For each message, the subclass is called to determine the effective size
-     *  of the message, and whether the aggregate batch size is within the range
-     *  accepted by the service.
-     */
-    protected List<LogMessage> buildBatch()
-    {
-        // presizing to a small-but-possible size to avoid repeated resizes
-        List<LogMessage> batch = new ArrayList<LogMessage>(512);
-
-        // we'll wait "forever" unless there's a shutdown timestamp in effect
-        LogMessage message = messageQueue.dequeue(shutdownTime);
-        if (message == null)
-            return batch;
-
-        long batchTimeout = System.currentTimeMillis() + config.batchDelay;
-        int batchBytes = 0;
-        int batchMsgs = 0;
-        while (message != null)
-        {
-            batchBytes += effectiveSize(message);
-            batchMsgs++;
-
-            // if this message would exceed the batch limits, push it back onto the queue
-            // the first message must never break this rule -- and shouldn't, as long as
-            // appender checks size
-            if (! withinServiceLimits(batchBytes, batchMsgs))
-            {
-                messageQueue.requeue(message);
-                break;
-            }
-
-            batch.add(message);
-            message = waitForMessage(batchTimeout);
-        }
-
-        return batch;
     }
 }