+++ aws-shared/src/main/java/com/kdgregory/logging/aws/sns/SNSLogWriter.java
@@ -116,6 +116,12 @@ extends AbstractLogWriter<SNSWriterConfig,SNSWriterStatistics,AmazonSNS>
     }
 
 
+    @Override
+    protected void stopAWSClient()
+    {
+        client.shutdown();
+    }
+
 //----------------------------------------------------------------------------
 //  Internal
 //----------------------------------------------------------------------------